Prevent users from adding overlapping markers
Show the user an error message if they try adding a marker that overlaps
with an existing one. To do that, pillage my PlexWeb repository to bring
in Overlay.js, which in turn brings in Animate, ConsoleLog, and Tooltip.
Those files have been ported more-or-less 1:1, so the only real changes
are in index.js.

This change also starts the process of keeping track of marker changes
client-side, so that we are always up to date and know when we may be
overlapping with an exiting marker.