Interact directly with the database for search
This is the start of an effort to avoid making calls to the web API,
since it's highly advised to shut down PMS when interacting directly
with the database. This first change uses the database directly to
search for shows. There's a downside in that  we aren't as forgiving as
Plex's search API that accounts for common spelling errors, but not
needing PMS running is more important.