Mark timestamps from the database as UTC-based
Timestamps are stored as UTC when adding or editing markers, but don't
indicate that they are when retrieved from the database. Append ' UTC'
to them so that the browser knows to correctly calculate the timezone
difference when displaying friendly dates. Note that I haven't verified
that Plex themselves add entries as UTC timestamps, but they should.