Add cache for marker breakdown
After the initial request for a marker breakdown, keep the counts in a
cache to return on subsequent requests to avoid making additional
potentially expensive queries to the database.

Also fix some bugs in the Animate class that caused animations to hang
if an element was removed and readded to the DOM in the middle of an
animation.