Create ThumbnailManager backend
Create the ThumbnailManager class that can read Plex's genereted preview
thumbnails and return a thumbnail closest to the desired timestamp. The
goal of the class is to provide better insight into where a marker is
being added, e.g. to find a good approximation for where end credits
start, avoiding the need to scrub through the episode itself in an
external program like VLC to get the right timestamps.

No frontend hookup yet, but thumbnails can be manually retrieved with
/t/metadataId/timestampInSeconds syntax.