Hook up ThumbnailManager to frontend
Add some basic UI to show thumbnails when adding or editing markers.

This change also includes a new (and sparse) settings dialog where users
can control the UI theme (in addition to the main toggle), and whether
they want to show thumbnails when adding/editing markers (in conjunction
with the server-side useThumbnails setting).