Abstract away the user configuration
Create a new PlexIntroEditorConfig class that handles reading
config.json, and provides read-only access to its values in app.js.

This is the hopeful start of an effort to reduce the number of
undocumented raw Objects used both server- and client-side.

Also restore the random deletion of a variable declaration that caused
the app to crash when retrieving the episodes in a season.