Add option for 'extended marker stats'
If `extendedMarkerStats` is enabled, cache all markers in the database
before launching the server. This process will create two complimentary
data structures:
* A map of marker tag ids to the marker data
* A hierarchy of section > show > season > episode > [...marker data],
 used for  aggregating marker statistics at the section->episode level.
Additionally, when a marker is added or deleted, update the cache so
that it stays up to date.

Nothing is hooked up client-side yet. This change just verifies the
backend behavior.