Use MarkerCacheManager for marker breakdowns
If extendedMarkerStats are enabled, use our already cached marker data
to calculate marker breakdowns (mapping the number of episodes that have
X number of markers). If extendedMarkerStats are disabled or for some
reason the cache manager fails to aggregate the data, fall back to the
existing process for collecting marker breakdowns.