More work toward surfacing 'extendedMarkerStats'
Send marker breakdown data at an arbitrary level to the client.
Instead of Marker*Node classes only holding a marker count, keep track
of the number of episodes in the 'level' that have `n` markers.

While this data is now sent to the client as `markerBreakdown`, it is
not yet integrated into the UI.