Convert everything to ES modules.
This allows for a common structure across server and client code, and
removes the need for gross hacks to get VSCode to bring in the right
imports for intellisense.

Also includes a major refactor of the ConsoleLog class, as its previous
self-referrential state was confusing (e.g. making references to an
external instance of itself and just assuming it exists).