Allow hiding thumbnails without disabling them.
Even if preview thumbnails are available, the user might not want to
always show them. Add a button when adding/editing a marker that will
show/hide the preview thumbnails, with a new setting that determines
whether thumbnails will start expanded or collapsed.

Also fix a couple tangential bugs/tweaks:
* More fallout from the appendChildren refactor that prevented us from
  styling marker table headers.
* Don't make a request for a preview thumbnail if we don't have a valid
  millisecond offset. Similarly fix a server-side bug that throws an
  exception if it gets an invalid thumbnail request.
* Make 'button' text unselectable, preventing it from being highlighted
  accidentally if it's clicked in quick succession.
* Add hover text to the settings dialog's Apply/Cancel buttons.
* Keep save/discard and options buttons top aligned, which prevents them
  from jumping around the table when ppreview thumbnails cause the row
  height to change.