Create PlexQueryManager to handle main SQL queries
This change creates a new PlexQueryManager class that takes ownership
of any direct references to the database in the main PlexIntroEditor
file. The actual processing of the query results remain behind, but the
queries themselves have been moved, in addition to the process of
creating the database connection and finding the persistent marker id.