Retrieve the most recently generated thumbnails
Some episodes with multiple versions will have multiple sd-index.bif
files. If that's the case, pick the most recent. Unfortunately markers
are associated with an episode as a whole, not a particular version of
one, and the most recent seems like a decent choice.

In a similar vein, adjust the getEpisodes query to grab the longest
episode for use in the Ctrl+Shift+E shortcut to enter the end of an
episode for credit markers, as otherwise we might be adding a marker
that's not actually at the end of the file, and Plex seems fine with
markers that go beyond the length of one version of the file.