Record manual marker actions to a new database
I recently discovered that Plex will wipe out all existing intro markers
for a season if a new episode is added to it, including any manually
added markers. This change attempts to start addressing this by
recording all marker actions into a separate database (if the feature is
enabled). Nothing is hooked up to the frontend yet, but some basic
manual validation has been done to ensure it's not wildly broken.

High level summary of changes:
* Create a MarkerBackupManager class that handles writing marker actions
  to a new database, as well as retrieving those entries.
* Add a setting that controls whether the backup database is enabled.
  This shouldn't be optional once it's been further validated, as
  enabling/disabling recording will lead to inaccurate results.
* Extend the default rows returned by the PlexQueryManager for marker
  queries. Namely, also include the season/show/section ids. Similarly,
  expand the common MarkerData type to include the new fields.
* Expand PlexQueryManager's addMarker to also handle reindexing of
  existing markers (to share more code between the basic 'add' case and
  the backup manager's 'restore' case).
* Add two new server endpoints that aren't hooked up client-side:
  * `purge_check` takes a metadataId and finds markers that exist in the
    backup database but not the Plex database
  * `restore` takes an old marker id from the backup database and
    restores it back to the Plex database.