Expand purged marker support
Allow the user to click the '(!)' text, which will show a table of all
purged markers for that episode. The user can then restore the lost
marker, or permanately ignore it.

Additional tweaks:
* Hackily allow a callback to be passed into Animate methods, to be
  called once an animation group has completed.
* Add a 'getIcon' method to ThemeColors to avoid constantly manually
  building an '/i/abcdef/icon.svg' string.

Additional notes:
While that core functionality works, it is still far from complete. The
biggest issue right now is that when a marker is restored/ignored, it
doesn't tell anyone. Marker counts aren't updated, and the cached
purge data isn't either. Refreshing the page is pretty much the only way
all necessary field will be updated in the UI.

Outside of those immediate issues, purged markers should also be
surfaced earlier, not only when looking at the episodes of a season.