Update marker cache if a show/season isn't found
When markers are loaded on startup, it's possible for the user to
(inadvisably) continue running PMS and add episodes. If we detect that a
new show or season has been added, attempt to update the marker cache.

This doesn't solve the case where a new episode is added to an existing
season, but could potentially be addressed in the future.