Add client-server 'find all purges' pipeline
Stand up some initial UI that will invoke a search for purged markers
markers in the current library. The 'Find Purged Markers' button is
currently hidden, but clicking it will initiate a request to the server
to retrieve a list of purged markers and show an overaly with the
result.

The requests are managed by a new PurgedMarkerManager. While currently
lacking in functionality, it will ideally become the sole endpoint for
purged marker management, incorporating/interfacing with the existing
PurgeTable class.