Add top-level purge detection
Instead of requiring a user to dig into a specific show to find purged
markers, add a button below the marker breakdown that will query for all
purged markers in the current section, launching an overlay that groups
all of them by show. This is a very large change, with the following key
pieces:

* On boot, if the backup manager and extended marker stats are enabled,
  build a map of all markers that have been purged from Plex's database.
* Adjust existing restore/ignore server-side methods to handle multiple
  markers at once.
* Delete PurgeTable.js and incorporate relevant bits into non-exposed
  classes in PurgedMarkerManager.js
* Create various new classes in PurgedMarkerManager.js to handle bulk
  restoration/ignoring of markers.
* Add a static singleton accessor to PlexClientState instead of needing
  to pass it around everywhere.
* Various additional logic and style tweaks to accomodate new scenarios.
* Enable the backup manager in the default config.json

This change also adds the bad '(!)' UI at the season level, more because
it was an in-progress change I didn't want to rip out before submitting
the new top-level mechanism, and I was too lazy to stash it/make a new
branch for it.