Ensure database queries are parameterized
For all fields that use input external to the class, ensure parametized
queries are used to avoid any potential sanitazation issues.

This change also tweaks some logging of the two database classes, adding
some additional logging and prefixing messages with the class name. It's
a tad annoying to add the prefix to each statement though, so I might
want to look into ways to get that information automatically.