Allow users to suspend the server without exiting
Instead of requringing a full shutdown, allow a suspended state where
the HTTP server is still running, but connections to the Plex database
are closed. This allows the user to suspend the server, restart Plex
without clashing with the intro editor, shutdown Plex again, and resume
the intro editor within the web page, avoiding the need to go back to
the commandline/IDE to restart editor.

The entrypoint is alongside the current restart/shutdown options within
the settings overlay. Some CSS was tweaked to better fit all three
options on narrow devices.

When in a suspended state the web app has no functionality, only able to
resume the server via an undismissible overlay.