Fix ButtonCreator input handler
The logic that handles 'enter == click' on buttons is flawed because
there are two distinct behaviors. If a 'thisArg' was provided to the
button creator then the callback will be called with two arguments, the
button itself and the event. However, if 'thisArg' isn't provided, then
the callback is invoked with a single argument, the event, and the
button will be 'this'. To reconcile this, create two separate callbacks,
with the single argument version forwarding to the two argument version,
passing in 'this' as the button.