Completely overhaul bulk restore purged markers
With the existing implementation, markers are restored in bulk by making
individual calls to addMarker. This can break easily if a single episode
has multiple markers that are being restored. With the right timing, we
can end up with multiple markers that have the same index because each
restoration doesn't know that the other is happening in parallel. This
in turn breaks the impercise getNewMarker method, which assumes unique
indexes, meaning when we go to add our newly restored markers to various
caches we can end up with multiple copies of a single marker, and no
copies of markers attached to the same episode.

The fix is to not rely on the existing addMarker, instead making a new
bulkRestore method that restores all markers in a single transaction and
is aware of multiple markers that belong to the same episode. Also
update getNewMarker to grab markers based on timestamps, not indexes.