Add 'pure mode' to avoid hacky db writes
Information about when a marker was last edited, and whether the
original marker was user or Plex generated is stored in the thumb_url
column of the taggings database, which intro markers don't otherwise
use. While in practice this seems harmless, there could be a future
where Plex decides to use this column, so it's probably a good idea to
allow users to disable storing data in this column, at the cost of
slightly reduced information when viewing existing markers.