Begin internalizing purged marker logic
Currently ResultRow classes do their own tracking of purged markers, but
it should be handled opaquely by the PurgedMarkerManager. This commit
lays the groundwork for that by creating the relevant classes within
PurgedMarkerManager to move all the retrieving/storing of purged markers
to be internal to the class, with accessors to retrieve purged markers
for a given show/season/episode. Nothing is hooked up to it yet, but it
works in theory.