Handle no-op restore of markers
In some edge-cases it's possible to try to bulk-restore markers but not
actually add any new markers to the database because identical markers
(with different ids) already exist. In this case we shouldn't fail the
operation, but instead assert that we hit this specific case and return
an empty array of restored markers, which the client already handles
gracefully.