Remove explicit portrait vs landscape CSS rules
Explicitly requiring a portrait aspect ratio for mobile view can break
if a tall keyboard changes a phone's detected orientation. Remove the
portrait check and only use max-wdith. That's not a perfect solution
either, and in general this app is terrible to use on mobile, but this
makes it fractionally better.