1.0.0 "release"
With the core functionality in place, a somewhat coherent UI, and decent
documentation, bump the package version to 1.0.0. In addition to that,
this change adjusts the help overlay to point to the wiki, and updates
README to give only a high level overview of the project, also pointing
to the wiki for additional details.