Integration tests proof of concept
Creates a framework for creating integration tests that run against the
intro editor using a test database and configuration file, as well as
some initial basic tests. Just like the first commits of the editor
itself, the test framework is very much a prototype that will need a lot
of cleanup if I continue down this road.

Mile-high hierarchy:
Test.js launches the TestRunner, which runs through all the TestClasses
that extend TestBase. TestBase runs through the methods of TestClass,
resetting the editor's state between methods. Once all TestClasses are
run, shut down the test server.

To get this to work, some changes were also required on the editor side:
* export getState so tests know the current status of the server
* Add argv parsing to detect test flags/config overrides
* Allow the Shutdown command to execute when the server is suspended
* Return the full MarkerData from editMarker instead of an "optimized"
  minimal version of it.
* Allow custom config paths in PlexIntroEditorConfig