Don't fail if data path can't be found
If preview thumbnails are disabled and a database path is provided, we
shouldn't care if we have a valid data path. However, we will still fail
to start if a data path isn't provided and we are unable to
automatically find its probable location. Avoid this by returning an
empty string instead of null from #getDefaultPlexDataPath, which won't
fail #getOrDefault, but will fail if we need a valid path in order to
build the database path or find the preview thumbnail path.