Don't load thumbnails if they're collapsed
Even if thumbnails are collapsed, as long as they're enabled we still
try to load them in the background. This is both a waste of resources,
and can also lead to unexpected behavior in the following situation:

1. Preview thumbnails enabled, not collapsed by default
2. Edit a marker, collapse thumbnails manually
3. Change the marker time and press Enter
4. Thumbnail appears again, even though it should be hidden

This change fixes that by not even attempting to grab a new thumbnail
when they're collapsed, only refreshing as soon as they're made visible
again.