Set focus to overlay button by default
Currently we don't change focus when launching a simple overlay, which
can lead to background elements retaining focus and causing unexpected
interactions. As one example, if the user has the 'Confirm Add' button
focused and presses enter, but the marker definition is invalid, an
overlay will appear with an 'OK' button. If I press 'Enter' again,
'Confirm Add' still has focus, so it just relaunches the same overlay.
Most people would expect the single overlay button to have focus
instead, so make that the default.