Add a Windows executable
v1.1.0 release, marked by the addition of a process for creating a
Windows exe. Now users can just extract a zip and double click instead
of installing node and MSVC build tools. This also involves adding
rollup to convert es6 modules to cjs, as nexe doesn't support modules
yet. Overall, the process is a hodgepodge of StackOverflow and other
forum suggestions that doesn't look great, but gets the job done.