Record and reindex markers after bulk action
I forgot two key pieces when implementing bulk shift and bulk delete -
recording the actions with BackupManager and updating the MarkerCache.
Implement both of those, as well as a new reindex method in the
PlexQueryManager to handle reindexing all markers under a given metadata
id, and use the new method in all relevant places, not just bulk
actions.

Also, return the array of deleted markers back to the client so client-
side caches can be updated properly as well. The client could use cached
data from showing the customize table, but it won't be available if the
user didn't show it, so it's easier to just return it directly.