Bulk Add Implementation
Large change that implements bulk add both client- and server-side. The
bulk add implementation is significantly longer than both shift and
delete, as it is much more customizable. There are three modes:
1. Fail - if there's overlap with any existing markers, fail the entire
   operation.
2. Ignore - if there's overlap with any existing markers, don't add the
   marker to the episode and continue the operation.
3. Merge - If there's overlap, merge the markers together. This is the
   bulk of the added complexity, as the merge can result in new, edited,
   and deleted markers all in the same operation, and UI client-side
   needs to know about all of it.

There are still many little quirks and inefficiencies here and there,
but the core functionality is reliable and the change is big enough
already.

In addition to the core implementation and basic tests, this change also
includes various other fixes/additions:

* Fix forceReset logic in editMarker
* Fix order of operation issues in notifyBulkActionChange
* Fix bulk delete of multiple markers from a single episode (UI only)
* Move flashButton to BulkActionCommon
* Allow tooltips to be removed from an element
* Small tweaks to serverMain handling
* await some async processes to avoid timing issues