Implement multiselect in bulk add overlay
Allow the typical Ctrl+Click/Shift+Click/Ctrl+Shift+Click behavior of
lists to allow multiple rows to be selected at once, and then checked/
unchecked all at once. While it currently only applies to the bulk-add
overlay, it's hopefully been designed in a way to allow for the other
overlays to adopt it fairly easily.

This change introduces:
* BulkActionRow - a base class that represents a row in a bulk action
  table.
* BulkActionTable - Holds the HTML table for a bulk action and handles
  the multiselect logic.