Add multiselect to the bulk shift overlay
Extract row creation/update out of BulkShiftOverlay and into a new class
that extends BulkActionRow and use BulkActionTable to display the rows
and enable multiselect.

Also adjust coloring slightly so it's easier to tell when a row is
selected, and also adjust bulk action buttons style to avoid shifting
on hover.