Check for a valid automatic data path on first run
On first run we check whether a user-supplied path for the data
directory and/or database exists, but not in the automatic case. There
are scenarios where we can't find the path automatically, or that path
doesn't exist, in which case we should ask the user to provide the full
path.