Initial ffmpeg-based thumbnail implementation
Initial implementation of using ffmpeg to extract precise thumbnails
from episodes instead of relying on Plex's generated ones. For now (and
for the forseeable future), the user has to have ffmpeg in their path in
order for this to work, otherwise we fall back to pre-generated thumbs
if they're available.

This also creates a cache of thumbnails on disk, so even if the
thumbnails are not in the in-memory cache, we don't have to regenerate
the same thumbnail over and over (until the cache is deleted).

Work that still needs to be done:
* Integrate into first-run experience.
* Have a system to delete cached thumbnails on disk, as right now they
  persist forever.
* Investigate Docker support.