Pre-cache purge data client-side
Grab purged data as soon as a library is selected if we know that things
cached server-side (backup actions and extended marker info enabled).

The main benefit is that we can immediately show the warning markers
when searching instead of waiting for 'Find Purged Markers' to be
clicked, or digging into a single season.