Limit movie result listings and search regex
Limit movie search results to 250 rows initially, but allow the user to
load everything if they really want to. Some people have ridiculously
large libraries, and we don't want to try loading 10K rows immediately.
This should probably do the same for TV shows, but it's much less likely
for someone to have 10K shows than it is for 10K movies.

On a search-related note, also add support for regex matching, which can
be helpful when e.g. looking for all movies that start with 'A'. Expose
the "real" sort/original titles to help with this, since it's necessary
for a case-sensitive search. On a similarly related note, expose the
original title of a movie if available in the search row.