Add additional handling to schema upgrade
When upgrading the backup database's schema from 3 to 4, one step is to
update any markers we've already added to the Plex database to use epoch
timestamps instead of time strings. However, if another version of
IntroEditor already ran that migration, our update query will fail. Add
some logic that checks whether the migration already ran, and if we
still can't get a valid value, reset it.