Include intro/credits info in marker breakdown
The existing marker breakdown only shows the number of episodes have a
total number of markers, it doesn't break it down by marker type. Make
that distinction with a questionable change to bit-shift credits counts
so that a single number can represent a unique (intro, credits) group.

This is one part of unlocking future work to allow filtering by whether
something has a specific type of marker.

Another part of this change is to share the MarkerBreakdown class so
there's a bit more structure to the client-side breakdown instead of raw
objects. This complicated the SerializedPlexType to PlexType conversion,
and while what this commit does works, it's be nice to make it cleaner.

As a part of the additional breakdown, add intro/credits info to the
extended marker stats tooltip, and add intro/credits specific breakdown
charts (and fix single-point pie charts).