Add support for basic filtering
Add a new row to the top of all sections that has a Filter button, which
launches an overlay that lets the user apply some basic filtering (based
on the number of intro/credits markers an item has). This new row will
potentially be expanded to contain additional options, like one to erase
all markers for a section, as suggested by #4.

The Filter UI currently reuses the Settings dialog style. Ideally, the
CSS names should be generalized, or the Filter CSS should be completely
separate.

Other adjustments made to better accommodate filtering:
* Add filter.svg, taken from PlexWeb
* When returning intro/credits breakdown buckets, ignore buckets that
  have no markers.
* Don't create any marker table HTML if we're lazy loading, do
  everything we we're actually initializing the table.
* Expose `markers()` from the MarkerTable
* When updating breakdowns after a purge restoration, also consider
  items in the section that aren't currently visible.