Fix broken movie marker tables
Due to the bad hierarchy between MovieResultRow, MovieData, and
MarkerTable, we can end up with a  MarkerTable that isn't attached to
the DOM, hiding it. Ensure the table is correctly reattached when a new
MovieResultRow is created. Also add some helpers to make it easier to
interact with a MarkerTable's visibility.