Add basic token-based search
Currently, search works by fuzzing the query (lowercase, remove
whitespace and special characters), then doing substring search on
fuzzed media titles. That works for most cases, but it can also be
helpful to do token-based search. This implementation is fairly simple,
doing the following:

1. When loading library data, break each title (and original title, if
   available) into individual words, and add the media item to a bucket
   for each word.
2. When searching, break the search query into individual words, and add
   the intersection of matching buckets to the search results (in
   addition to the regular fuzzy search results).

This allows for something like 'Ring Lord' to match 'The Lord of the
Rings', when it previously wouldn't because 'ringlord' isn't a substring
'thelordoftherings'.

Perf and memory is a potential concern, but basic tests don't show any
noticeable impact. My libraries are small compared to some users though.