Make marker edit Enter shortcuts trigger on keyup
When 'Enter' is captured as a keydown event and used to commit a
failable action, we might show error UI that sets focus to an enterable
element before the subsequent keyup event is fired, which our new UI
then captures. To avoid this, ensure cases like this capture the event
on keyup to avoid unwanted keydown/keyup propagation.