Fix rounding shortcuts
Rounding shortcuts (`\` and `Shift+\`) are broken when `Alt` is pressed,
because it always multiplies the time difference by 5, which means
attempting to round something like "1.123" to the nearest second results
in a change of 0.615 (0.123 * 5) seconds instead of 0.123. Fix that, and
instead make Alt round to the nearest 5 or 0.5, depending on `Shift`.