Ensure multiselect checkboxes are always visible
When customizing bulk actions, the checkboxes used to check/uncheck
multiple selected items are aligned with the first selected row, meaning
they might go off the screen. Fix that by ensuring the checkboxes are
pinned to either the top or bottom of the window, depending on where the
first selected row is in relation to the current viewport.

In addition to that, this change also includes the following:
* Add a label indicating how many items are currently selected
* Wrap multi-select elements in their own div so only a single element
  needs to be positioned absolutely.
* Add onDismiss callbacks to Overlay. This was added to aid with
  ensuring that window scroll listeners used to reposition the
  multiselect checkboxes are properly removed once the customization
  table is removed.
* Add isShowing/get methods to Overlay to avoid reaching into the
  "internals" of it by querying for `#mainOverlay` directly.