Remove Plex DB hack commandeering thumb_url
When a marker is added or edited, the thumb_url column of the taggings
table is used to store information about when the marker was added/
edited, as well as whether the marker was created by the user. This
design choice predates the backup database, which also contains all of
this information, and is much riskier in the event where Plex themselves
decide to make use of thumb_url for markers. Remove this hack by doing
the following:

* When initializing the marker backup manager, populate a cache of
  userCreated/modifiedAt data. This isn't strictly necessary, but is a
  performance optimization that lets us avoid constantly querying the
  backup database.
* When a marker is added or edited, ensure the cache above is correctly
  updated to reflect the latest state. This is a bit disjointed right
  now, and seems a bit fragile, but it works.
* Instead of writing to/reading from thumb_url, ask the cache above for
  the modified date/created by user flag.

Tangential changes made in response to the core changes above:

* Remove pureMode setting, as it no longer does anything.
* Remove backupAction setting, as its now always enabled. It's very
  helpful, and the only downside would be a bug that prevents other
  features from working, and nothing's been reported in the year+ it's
  been a feature.
* Tweak date added tooltip to account for new standards.
* Fix bug that prevented purged markers from indicating whether they
  were user created.
*