Add loading animation when clicking on a season
With ffmpeg thumbnails enabled, clicking on a season makes a call to
check whether all episode files exist, which can be slow depending on
where the files are stored. Add a loading icon after clicking on a
season to let the user know something's happening.

The real fix is to make the thumbnail check async, or call it when the
episode itself is clicked and block it there, but this is a decent
workaround that is helpful in the case of a slow server anyway.