Remove Animate.js
Following up on the previous change to use CSS animations in Overlay.js
instead of my homegrown solution, this change replaces all uses of
Animate.js with built-in animations. The result is a markedly better
experience for all transitions (even if not everything translated 1:1).
The difference is especially noticeable on HFR monitors, as my old
hacked-together animation library limited transitions to 60 FPS.

Core change:
* Replace all Animation calls with ele.animate
* Create AnimationHelpers.js with (awaitable) helpers for common
  animation scenarios.
* Delete Animate.js

A few other tangential changes were made as well:
* Better purge table row removal animation - work around tr height
  limitations to allow a smooth slide up when a row is removed.
* Add getHex helper to ThemeColors to retrieve a full hex color string
  (with optional opacity).
* Fix a bug that prevented PurgedMarkerManager's "ignored failed"
  animation from triggering.
* Rework #clearOverlayAfterPurge to make use of Overlay's new
  replace-in-place support.