Fix unexpected key press chaining
ButtonCreator attaches a keyup listener to all elements, triggering a
'click' when Enter is pressed. However, because regular inputs trigger
on keydown, if a regular input keydown launches an Overlay that changes
the focus to a Button, on keyup we'll then immediately invoke that
button action, even though the initial key press was associated with a
different element. Use keydown instead to fix this, and align with
normal inputs.

The initial motivation behind the keyup event was to avoid potential
issues around users holding a key down for too long and triggering
multiple clicks, but that can be handled by checking the repeat property
of the event.