Icon Overhaul
Switch to inline SVG elements that use currentColor for color styling
instead of server requests for specific colors.

There were two main hurdles I ran into:
* We aren't guaranteed browser-handled caching like <img> sources. A
  naive implementation might result in asking the server for 200 copies
  'arrow.svg' at once.
* Fetching the icon is async if it's not already cached, and <img>'s
  built-in mechanism of handling this isn't available.

For #1, create our own caching mechanism. When an icon request comes in:
* Check if it's cached. If it is, do a deep clone and return it.
* If it isn't, request it from the server. Wrap that request in a
  Promise so that if a request for the same icon comes in before the
  first request finishes, we can await on that promise and not make
  multiple requests.
* Once the request returns, cache it and return a deep clone of it.

For #2, keep everything ~synchronous by returning a placeholder SVG if
the icon isn't cached yet. Once the icon is retrieved, replace it with
the real icon, all "behind-the-scenes" in SVGHelper. The downside to
this approach is that it isn't necessarily safe to manipulate the
element as soon as its requested, as it might be the soon-to-be-replaced
placeholder. Another approach would be to make the retrieval async, but
I didn't like the idea of making the caller remember when to `await`
when building UI. Remembering to wrap icons in some other element and
applying listeners/attributes to that seemed easier to keep track of.