Allow negative timestamps in Bulk Add
Treat negative timestamps as an offset from the end of an item, so e.g.
-1:00 to -30000 is "one minute before the end of the media until 30
seconds before it ends." Fairly straightforward change due to the
infra changes necessary to add markers based on chapters, with the only
tricky bit being the proper handling of `-0` ("end of the media").

Also return a more descriptive error message when something fails, and
add a tooltip when an invalid timespan is entered for an episode.