Use CSS themes for light/dark mode differences
CSS is still an absolute mess, but this gets things slightly more
consistent state. The core change is to create CSS variables for all
light/dark color differences, shrinking the light/dark specific
stylesheets and providing a starting point for a hopeful consolidation
of theme colors.

Some styling was tweaked along the way, but the overall theme remains
the same.