Reduce server queries when getting movie info
When expanding a movie row for the first time, we make up to 4 queries:
1. Get marker data
2. Get thumbnail info
3. Check for purge data
4. Get chapter info

Add a new endpoint to combine 1, 2, and 4 into a single server query,
and fix an issue that would cause us to check for purges of items where
we know the section already has complete data.

Also fix a bug that caused us to add movies to a PurgedTVSection instead
of a PurgedMovieSection.