Better handling of multiple animations
Because animating height doesn't work without explicit bounds, the
animation helpers temporarily set an explicit height. However, if
another height-altering animation starts before the original one ends,
we'll incorrectly assume that the temporary height is the one we want to
restore at the end, which isn't the case. Set attributes outside of the
element's style to save the real value we want to restore.