Add pathMappings option for FFmpeg thumbnails
FFmpeg thumbnails relies on the file path set in the database, but if
Marker Editor is being run in a different environment, the paths might
not line up with the current filesystem. Allow users to specify simple
prefix replacements in the form of `{ from: [str], to: [str] }` entries in
config.json, and integrate it into the FirstRunConfig flow.

Then, when we then check whether a media file exists, run the path
through the mappings and make any necessary replacements, still falling
back to the original path if none of the modified paths exist.

Tangential changes:
* FirstRunConfig
  * Better messaging around thumbnail types, and determining whether a
    specific thumbnail type can be used.
  * Allow users to skip specifying the data path, as it's not always
    required, and may not exist.
* ThumbnailManager
  * Switch to fs/promises in ThumbnailManager, which can help with
    concurrency (e.g. checking whether every episode in a season exists)
  * Create ThumbnailNotGeneratedError to avoid hacky method of
    determining whether our ffmpeg command failed.