Remove MarkerTableRow's dependency on ResultRow
This one is a bit superficial - if some spurious validation is removed, we can
get away with a JSDoc typedef import to keep intellisense and remove
dependencies on ResultRow.js.

Also turn MediaItemWithMarkerTable into a @typedef, since it's not something
that's ever actually instantiated.

This change reduces the number of circular dependencies from 3 to 2.