Improve first load
Improve load experience. Benefits are relatively minimal since it's not expected
that the user frequently reloads the main page, but anecdotally the reload
experience is much better than before (at the expense of significantly more dev
dependencies, but I'm starting to realize that doesn't really matter in the
grand scheme of things).

* Move core CSS to index.html, removing allowing it to be applied during the
  initial request.
  * Add minimal theme parsing logic to index.html so the initial paint uses the
    system- or user-set theme.
* Use webpack to consolidate client code into a single request. Without webpack,
  there's a chain of 5+ requests required to load the client code, which is
  noticeably slower than loading the minified bundle.
* Convert build.cjs to a module (mjs)